+++ lib/Auth/Process/IsCesnetEligible.php
@@ -1,16 +1,24 @@
 <?php
 
+namespace SimpleSAML\Module\cesnet\Auth\Process;
+
+use SimpleSAML\Module\perun\LdapConnector;
+use SimpleSAML\Module\perun\RpcConnector;
+use SimpleSAML\Module\perun\AdapterLdap;
+use SimpleSAML\Module\perun\AdapterRpc;
+use SimpleSAML\Logger;
+use SimpleSAML\Error\Exception;
+
 /**
- * Class sspmod_cesnet_Auth_Process_IsCesnetEligible
+ * Class IsCesnetEligible
  *
  * This class put the timestamp of last login into list of Attributes, when at least one value of attribute
  * 'eduPersonScopedAffiliation' is marked as isCesnetEligible in CESNET LDAP
  *
  * @author Pavel Vyskocil <vyskocilpavel@muni.cz>
  */
-class sspmod_cesnet_Auth_Process_IsCesnetEligible extends SimpleSAML_Auth_ProcessingFilter
+class IsCesnetEligible extends \SimpleSAML\Auth\ProcessingFilter
 {
-
     const CONFIG_FILE_NAME = 'module_cesnet_IsCesnetEligible.php';
     const ORGANIZATION_LDAP_BASE = 'ou=Organizations,o=eduID.cz,o=apps,dc=cesnet,dc=cz';
 
@@ -28,12 +36,12 @@ class sspmod_cesnet_Auth_Process_IsCesnetEligible extends SimpleSAML_Auth_Proces
     private $eduPersonScopedAffiliation = array();
 
     /**
-     * @var sspmod_perun_LdapConnector
+     * @var LdapConnector
      */
     private $cesnetLdapConnector;
 
     /**
-     * @var sspmod_perun_RpcConnector
+     * @var RpcConnector
      */
     private $rpcConnector;
 
@@ -42,7 +50,7 @@ class sspmod_cesnet_Auth_Process_IsCesnetEligible extends SimpleSAML_Auth_Proces
         parent::__construct($config, $reserved);
 
         if (!isset($config[self::CESNET_ELIGIBLE_LAST_SEEN_ATTR])) {
-            throw new SimpleSAML_Error_Exception(
+            throw new Exception(
                 "cesnet:IsCesnetEligible - missing mandatory configuration option '" .
                 self::CESNET_ELIGIBLE_LAST_SEEN_ATTR . "'."
             );
@@ -56,8 +64,8 @@ class sspmod_cesnet_Auth_Process_IsCesnetEligible extends SimpleSAML_Auth_Proces
 
         $this->cesnetEligibleLastSeenAttr = $config[self::CESNET_ELIGIBLE_LAST_SEEN_ATTR];
 
-        $this->cesnetLdapConnector = (new sspmod_perun_AdapterLdap(self::CONFIG_FILE_NAME))->getConnector();
-        $this->rpcConnector = (new sspmod_perun_AdapterRpc())->getConnector();
+        $this->cesnetLdapConnector = (new AdapterLdap(self::CONFIG_FILE_NAME))->getConnector();
+        $this->rpcConnector = (new AdapterRpc())->getConnector();
     }
 
     public function process(&$request)
@@ -67,7 +75,7 @@ class sspmod_cesnet_Auth_Process_IsCesnetEligible extends SimpleSAML_Auth_Proces
         if (isset($request['perun']) && isset($request['perun']['user'])) {
             $user = $request['perun']['user'];
         } else {
-            SimpleSAML\Logger::debug(
+            Logger::debug(
                 "cesnet:IsCesnetEligible - " .
                 "Request doesn't contain User, so attribute 'isCesnetEligible' won't be stored."
             );
@@ -77,40 +85,47 @@ class sspmod_cesnet_Auth_Process_IsCesnetEligible extends SimpleSAML_Auth_Proces
         $this->idpEntityId = $request['saml:sp:IdP'];
 
         if (isset($request['Attributes']['eduPersonScopedAffiliation'])) {
-            $this->eduPersonScopedAffiliation = $request['Attributes']['eduPersonScopedAffiliation'];
+            $this->eduPersonScopedAffiliation
+                = $request['Attributes']['eduPersonScopedAffiliation'];
         } else {
-            SimpleSAML\Logger::error(
+            Logger::error(
                 "cesnet:IsCesnetEligible - Attribute with name 'eduPersonScopedAffiliation' did not received from IdP!"
             );
         }
 
         $isHostelVerified = false;
-        if ($request['saml:sp:IdP'] === self::HOSTEL_ENTITY_ID && isset($request['Attributes']['loa'])
-            && $request['Attributes']['loa'][0] == 2) {
+        if ($request['saml:sp:IdP'] === self::HOSTEL_ENTITY_ID &&
+            isset($request['Attributes']['loa'])
+            && $request['Attributes']['loa'][0] == 2
+        ) {
             $isHostelVerified = true;
-            SimpleSAML\Logger::debug("cesnet:IsCesnetEligible - The user was verified by Hostel.");
+            Logger::debug("cesnet:IsCesnetEligible - The user was verified by Hostel.");
         }
 
         try {
             if (!empty($user)) {
-                $this->cesnetEligibleLastSeen = $this->rpcConnector->get('attributesManager', 'getAttribute', array(
-                    'user' => $user->getId(),
-                    'attributeName' => $this->cesnetEligibleLastSeenAttr,
-                ));
+                $this->cesnetEligibleLastSeen = $this->rpcConnector->get(
+                    'attributesManager',
+                    'getAttribute',
+                    array('user' => $user->getId(), 'attributeName' => $this->cesnetEligibleLastSeenAttr,)
+                );
             }
 
-            if ((!empty($this->eduPersonScopedAffiliation) && $this->isCesnetEligible()) || $isHostelVerified) {
+            if ((!empty($this->eduPersonScopedAffiliation) && $this->isCesnetEligible())
+                || $isHostelVerified
+            ) {
                 $this->cesnetEligibleLastSeen['value'] = date("Y-m-d H:i:s");
 
                 if (!empty($user)) {
-                    $this->rpcConnector->post('attributesManager', 'setAttribute', array(
-                        'user' => $user->getId(),
-                        'attribute' => $this->cesnetEligibleLastSeen,
-                    ));
+                    $this->rpcConnector->post(
+                        'attributesManager',
+                        'setAttribute',
+                        array('user' => $user->getId(), 'attribute' => $this->cesnetEligibleLastSeen,)
+                    );
                 }
             }
         } catch (Exception $ex) {
-            SimpleSAML\Logger::warning("cesnet:IsCesnetEligible - " . $ex->getMessage());
+            Logger::warning("cesnet:IsCesnetEligible - " . $ex->getMessage());
         }
 
         if ($this->cesnetEligibleLastSeen['value'] != null) {
@@ -124,12 +139,14 @@ class sspmod_cesnet_Auth_Process_IsCesnetEligible extends SimpleSAML_Auth_Proces
      */
     private function isCesnetEligible()
     {
-
-        $allowedAffiliations = $this->getAllowedAffiliations($this->idpEntityId);
+        $allowedAffiliations
+            = $this->getAllowedAffiliations($this->idpEntityId);
         foreach ($this->eduPersonScopedAffiliation as $userAffiliation) {
             $userAffiliationWithoutScope = explode("@", $userAffiliation)[0];
-            if (!is_null($userAffiliationWithoutScope) && !empty($userAffiliationWithoutScope)
-                && in_array($userAffiliationWithoutScope, $allowedAffiliations)) {
+            if (!is_null($userAffiliationWithoutScope) &&
+                !empty($userAffiliationWithoutScope) &&
+                in_array($userAffiliationWithoutScope, $allowedAffiliations)
+            ) {
                 return true;
             }
         }
@@ -145,21 +162,25 @@ class sspmod_cesnet_Auth_Process_IsCesnetEligible extends SimpleSAML_Auth_Proces
     {
         $allowedAffiliations = array();
 
-		try {
-			$affiliations = $this->cesnetLdapConnector->searchForEntity(self::ORGANIZATION_LDAP_BASE,'(entityIDofIdP=' . $idpEntityId . ')', array(
-				'cesnetcustomeraffiliation'))['cesnetcustomeraffiliation'];
+        try {
+            $affiliations = $this->cesnetLdapConnector->searchForEntity(
+                self::ORGANIZATION_LDAP_BASE,
+                '(entityIDofIdP=' . $idpEntityId . ')',
+                array('cesnetcustomeraffiliation')
+            )['cesnetcustomeraffiliation'];
 
-			if (empty($affiliations)) {
-				SimpleSAML\Logger::debug("cesnet:IsCesnetEligible - Received empty response from LDAP, entityId " . $idpEntityId . " was probably not found.");
-			} else {
-				foreach ($affiliations as $affiliation) {
-					array_push($allowedAffiliations, $affiliation);
-				}
-			}
-		} catch (Exception $ex) {
-			SimpleSAML\Logger::warning("cesnet:IsCesnetEligible - Unable to connect to LDAP!");
-		}
+            if (empty($affiliations)) {
+                Logger::debug("cesnet:IsCesnetEligible - Received empty response from LDAP, entityId "
+                    . $idpEntityId . " was probably not found.");
+            } else {
+                foreach ($affiliations as $affiliation) {
+                    array_push($allowedAffiliations, $affiliation);
+                }
+            }
+        } catch (Exception $ex) {
+            Logger::warning("cesnet:IsCesnetEligible - Unable to connect to LDAP!");
+        }
 
-		return $allowedAffiliations;
-	}
+        return $allowedAffiliations;
+    }
 }