+++ themes/cesnet/consent/consentform.php
@@ -1,4 +1,8 @@
 <?php
+
+use SimpleSAML\Module;
+use SimpleSAML\Utils\Arrays;
+
 /**
  * Template form for giving consent.
  *
@@ -57,15 +61,11 @@ $attributes = $this->data['attributes'];
 $this->data['header'] = $this->t('{consent:consent:consent_header}');
 
 $this->data['head'] = '<link rel="stylesheet" media="screen" type="text/css" href="' .
-    SimpleSAML\Module::getModuleUrl('consent/style.css') . '" />';
+    Module::getModuleUrl('consent/assets/css/consent.css') . '" />';
 $this->data['head'] .= '<link rel="stylesheet" media="screen" type="text/css" href="' .
-    SimpleSAML\Module::getModuleUrl('cesnet/res/css/consent.css') . '" />';
+    Module::getModuleUrl('cesnet/res/css/consent.css') . '" />';
 
 $this->includeAtTemplateBase('includes/header.php');
-?>
-
-
-<?php
 
 if (array_key_exists('descr_purpose', $this->data['dstMetadata'])) {
     echo '</p><p>' . $this->t(
@@ -73,17 +73,15 @@ if (array_key_exists('descr_purpose', $this->data['dstMetadata'])) {
         array(
             'SPNAME' => $dstName,
             'SPDESC' => $this->getTranslation(
-                SimpleSAML\Utils\Arrays::arrayize(
+                Arrays::arrayize(
                     $this->data['dstMetadata']['descr_purpose'],
                     'en'
                 )
             ),
-            )
+        )
     );
 }
-?>
 
-<?php
 if ($this->data['sppp'] !== false) {
     echo "<p>" . htmlspecialchars($this->t('{consent:consent:consent_privacypolicy}')) . " ";
     echo "<a target='_blank' href='" . htmlspecialchars($this->data['sppp']) . "'>" . $dstName . "</a>";
@@ -103,7 +101,6 @@ echo present_attributes($this, $attributes, '');
 
     <div class="row">
         <div class="col-xs-6">
-
 
             <form action="<?php echo htmlspecialchars($this->data['yesTarget']); ?>">
                 <?php
@@ -129,13 +126,10 @@ echo present_attributes($this, $attributes, '');
                     <?php echo htmlspecialchars($this->t('{consent:consent:yes}')) ?>
                 </button>
 
-
             </form>
-
 
         </div>
         <div class="col-xs-6">
-
 
             <form action="<?php echo htmlspecialchars($this->data['noTarget']); ?>">
 
@@ -148,11 +142,12 @@ echo present_attributes($this, $attributes, '');
                 <button type="submit" class="btn btn-lg btn-default btn-block  btn-no" name="no" id="nobutton">
                     <?php echo htmlspecialchars($this->t('{consent:consent:no}')) ?>
                 </button>
-            </form>
 
+            </form>
 
         </div>
     </div>
+
 <?php
 
 $this->includeAtTemplateBase('includes/footer.php');