+++ themes/cesnet/consent/noconsent.php
@@ -1,5 +1,7 @@
 <?php
 
+use SimpleSAML\Module;
+
 if (array_key_exists('name', $this->data['dstMetadata'])) {
     $dstName = $this->data['dstMetadata']['name'];
 } elseif (array_key_exists('OrganizationDisplayName', $this->data['dstMetadata'])) {
@@ -7,21 +9,21 @@ if (array_key_exists('name', $this->data['dstMetadata'])) {
 } else {
     $dstName = $this->data['dstMetadata']['entityid'];
 }
+
 if (is_array($dstName)) {
     $dstName = $this->t($dstName);
 }
-$dstName = htmlspecialchars($dstName);
 
+$dstName = htmlspecialchars($dstName);
 
 $this->data['header'] = $this->t('{consent:consent:noconsent_title}');
 
 $this->data['head'] = '<link rel="stylesheet" media="screen" type="text/css" ' .
     'href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" />';
 $this->data['head'] .= '<link rel="stylesheet" media="screen" type="text/css" ' .
-    'href="' . SimpleSAML\Module::getModuleUrl('cesnet/res/css/consent.css') . '" />';
+    'href="' . Module::getModuleUrl('cesnet/res/css/consent.css') . '" />';
 
 $this->includeAtTemplateBase('includes/header.php');
-
 
 echo '<p>' . $this->t('{consent:consent:noconsent_text}', array('SPNAME' => $dstName)) . '</p>';
 
@@ -43,6 +45,5 @@ echo('<p><a class="btn btn-default" href="' . htmlspecialchars($this->data['logo
 echo('<i class="glyphicon glyphicon-ban-circle"></i> ');
 echo($this->t('{consent:consent:abort}', array('SPNAME' => $dstName)));
 echo('</a></p>');
-
 
 $this->includeAtTemplateBase('includes/footer.php');