+++ themes/cesnet/default/errorreport.php
@@ -1,6 +1,8 @@
 <?php
+
 $this->data['header'] = $this->t('errorreport_header');
 $this->includeAtTemplateBase('includes/header.php');
+
 ?>
 
     <p><?php echo $this->t('errorreport_text'); ?></p>